requirements.txt
asgiref==3.5.2
certifi==2022.9.24
decorator==5.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gitdb==4.0.10
GitPython==3.1.29
ipython==8.7.0
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytils==0.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
